운동 시작, 업데이트, 종료 로직 추가

+++ lib/platform/models/user_state_model.dart
@@ -6,8 +6,8 @@ part 'user_state_model.g.dart';
 class UserStateModel {
   int id;
   int userId;
-  int stamina;
-  int dailyGoReward;
+  double stamina;
+  double dailyGoReward;
   String createdBy;
   String createdDate;
   String lastModifiedBy;
